feat: migrate to async OkHttp enqueue with callbacks and image caching

+++ app/src/main/java/ru/netology/nmedia/util/Utils.kt
@@ -2,7 +2,7 @@ package ru.netology.nmedia.util
 
 import kotlin.math.floor
 
-fun formatCount(count: Long): String {
+fun formatCount(count: Int): String {
     return when {
         count < 1_000 -> count.toString()
         count < 10_000 -> {
@@ -16,4 +16,3 @@ fun formatCount(count: Long): String {
         }
     }
 }
-
